pyproject.toml
[project]
name = "st-api-test"
version = "0.1.0"
description = "Samsung SmartThings REST API 테스트 및 LLM Tool 패키지"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api_explorer", "smartthings_tools"]
